Reorganize and improve test coverage
